+++ app/helpers/notifications_helper.rb
@@ -1,0 +1,5 @@
+module NotificationsHelper
+  def unchecked_notifications
+    @notifications = current_user.passive_notifications.where(checked: false)
+  end
+end
