+++ app/views/layouts/_header.html.erb
@@ -1,4 +1,4 @@
-<header class="w-full px-4 sm:px-10 py-2">
+<header class="w-full px-4 sm:px-10 py-2 " data-controller="notifications">
   <!-- PC時 (1段表示) -->
   <div class="hidden md:flex items-center justify-between w-full h-16">
     <!-- 左側：ロゴ -->
@@ -39,6 +39,18 @@
               <span class="material-symbols-outlined text-gray-600 flex items-center justify-center w-10 h-10" style="font-size: 40px;">account_circle</span>
           <% end %>
         <% end %>
+      
+         <!-- 通知アイコン -->
+        <div class="relative">
+          <%= link_to notifications_path, class: "flex items-center justify-center relative" do %>
+            <span class="material-symbols-outlined  text-gray-600 text-3xl">notifications</span>
+            <% if unchecked_notifications.any? %>
+              <span class="absolute top-0 right-0 transform translate-x-2 -translate-y-2 bg-red-500 text-white text-xs font-bold w-5 h-5 flex items-center justify-center rounded-full">
+                <%= unchecked_notifications.count %>
+              </span>
+            <% end %>
+          <% end %>
+        </div>
       <% else %>
         <%= link_to new_user_session_path, class: "flex items-center justify-center", title: "ログイン" do %>
           <span class="material-icons text-gray-600 text-3xl">login</span>
