+++ app/views/notifications/index.html.erb
@@ -1,0 +1,41 @@
+<div class="container mx-auto px-4 py-6">
+  <h2 class="text-2xl font-bold text-gray-800 mb-4">通知一覧</h2>
+
+  <div class="bg-white shadow-md rounded-lg p-4">
+    <% if @unchecked_notifications.any? || @checked_notifications.any? %>
+      <% (@unchecked_notifications + @checked_notifications).each do |notification| %>
+        <div class="notification-item flex items-center justify-between border-b border-gray-300 py-2 px-3">
+          <div class="flex items-center space-x-3">
+            <% if notification.visitor.avatar.attached? %>
+              <%= image_tag notification.visitor.avatar.variant(resize: "40x40"), class: "w-10 h-10 rounded-full" %>
+            <% else %>
+              <span class="material-icons text-gray-500 text-4xl">account_circle</span>
+            <% end %>
+            <div>
+              <% if notification.action == 'favorite_review' %>
+                <p class="text-gray-700">
+                  <%= link_to notification.visitor.name, profile_path(notification.visitor), class: "font-bold text-blue-600 hover:underline" %>
+                  さんがあなたの投稿に
+                  <%= link_to "いいねしました", review_path(notification.review), class: "text-blue-600 hover:underline" %>
+                </p>
+              <% elsif notification.action == 'comment' %>
+                <p class="text-gray-700">
+                  <%= link_to notification.visitor.name, profile_path(notification.visitor), class: "font-bold text-blue-600 hover:underline" %>
+                  さんがあなたの投稿に
+                  <%= link_to "コメントしました", review_path(notification.review), class: "text-blue-600 hover:underline" %>
+                </p>
+              <% end %>
+              <p class="text-sm text-gray-500"><%= time_ago_in_words(notification.created_at) %>前</p>
+            </div>
+          </div>
+        </div>
+      <% end %>
+    <% else %>
+      <p class="text-center text-gray-600">通知はありません。</p>
+    <% end %>
+  </div>
+
+  <div class="text-center mt-4">
+    <%= link_to "戻る", root_path, class: "inline-block bg-gray-500 hover:bg-gray-600 text-white text-sm font-semibold px-4 py-2 rounded-md shadow-md transition duration-200" %>
+  </div>
+</div>
