+++ config/routes.rb
@@ -24,6 +24,8 @@ Rails.application.routes.draw do
 
   resources :items, only: [ :index, :show ]
 
+  resources :notifications, only: [:index]
+
   get "/terms_of_service", to: "static_pages#terms_of_service"
   get "/privacy_policy", to: "static_pages#privacy_policy"
